# Finding the maximum performance