https://github.com/hatoo/oha/issues/617#issuecomment-2560991130

Only for a fixed number of requests run.

```
❯ TOKIO_WORKER_THREADS=16 oha -n 4000000 -c 1000 --no-tui http://localhost:3000
Summary:
  Success rate: 100.00%
  Total:        6.2924 secs
  Slowest:      0.0923 secs
  Fastest:      0.0000 secs
  Average:      0.0016 secs
  Requests/sec: 635684.8490

  Total data:   49.59 MiB
  Size/request: 13 B
  Size/sec:     7.88 MiB
```

```
❯ TOKIO_WORKER_THREADS=16 cargo run --release -- -n 4000000 -c 1000 --no-tui http://localhost:3000
Summary:
  Success rate: 100.00%
  Total:        5.7571 secs
  Slowest:      0.1873 secs
  Fastest:      0.0000 secs
  Average:      0.0014 secs
  Requests/sec: 694790.7464

  Total data:   49.59 MiB
  Size/request: 13 B
  Size/sec:     8.61 MiB
```

```
❯ wrk -t 16 -c 1000  http://localhost:3000
Running 10s test @ http://localhost:3000
  16 threads and 1000 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.23ms    1.19ms  21.26ms   90.28%
    Req/Sec    56.31k     9.45k  160.94k    78.17%
  9026938 requests in 10.05s, 1.09GB read
Requests/sec: 898204.31
Transfer/sec:    111.36MB
```

note: Set TOKIO_WORKER_THREADS to the number of actual physical cpus (no hyper-threading) actually helps performance.
It's another issue.