@hatoo 2025-01-06
Not sending to channel per requests greatly improve performance

note: `--profile pgo` doesn't mean it uses PGO, just profile for PGO build (enabling lto etc.)

```
❯ cargo run --profile pgo -- -n 6000000 -c 1000 --no-tui http://localhost:3000
Summary:
  Success rate: 100.00%
  Total:        6.9893 secs
  Slowest:      0.2570 secs
  Fastest:      0.0000 secs
  Average:      0.0011 secs
  Requests/sec: 858459.4633

  Total data:   74.39 MiB
  Size/request: 13 B
  Size/sec:     10.64 MiB
```

```
❯ wrk -t16 -c 1000 http://localhost:3000
Running 10s test @ http://localhost:3000
  16 threads and 1000 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.19ms    1.15ms  19.68ms   90.30%
    Req/Sec    58.90k     8.17k  177.76k    78.10%
  9435690 requests in 10.06s, 1.14GB read
Requests/sec: 938274.13
Transfer/sec:    116.33MB
```